requirements.txt
Flask~=3.0.3
colorama~=0.4.6
waitress~=3.0.0
pyserial~=3.5